ai-daily-assistant/requirements.txt
slack-bolt>=1.18.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.22
sentence-transformers>=2.2.2
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0.1
structlog>=23.2.0
python-dotenv>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
streamlit>=1.31.0
